evaluation/RealWorldQA/requirements.txt
vllm
transformers
qwen_vl_utils
pandas
numpy
Pillow
tqdm
requests
validators
torch
torchvision
accelerate
sentencepiece
flash_attn